Reorganize repo: move old files to legacy, add clean configs
- Move 38 old configs to legacy/old_configs/
- Move old scripts, docs, slurm jobs to legacy/
- Create clean model configs in configs/models/ (5 gemma3 IT models)
- Update base.yaml with current defaults
- Simplify README.md, move pipeline diagram to src/README.md
- All models now configured for 1 GPU, 40GB

+++ src/steering/apply.py
@@ -46,7 +46,7 @@ def steering_hook(
 
     def hook(_module, _input, output):
         hidden = output[0] if isinstance(output, tuple) else output
-        steer = scale * vector.to(hidden.device)
+        steer = scale * vector.to(hidden.device, dtype=hidden.dtype)
         steer = steer.unsqueeze(0).unsqueeze(0)  # broadcast over batch and seq
         hidden = hidden + steer
         if isinstance(output, tuple):
@@ -58,4 +58,3 @@ def steering_hook(
         yield handle
     finally:
         handle.remove()
-
